+++ src/lib/st-two-list-selection/list-selection/list-selection.component.ts
@@ -28,9 +28,12 @@ export class ListSelectionComponent {
    @Input() @StRequired() searchPlaceholder: string;
    @Input() @StRequired() qaTag: string;
    @Input() important: boolean = false;
+   @Input() hasAllList?: boolean = false;
    @Input() hasSearch: boolean = true;
+   @Input() itemAll?: StTwoListSelectionElement;
    @Input() orderOptions: StDropDownMenuItem[] = [];
    @Input() mode: 'compact' | 'normal' = 'normal';
+   @Input() showSearchNumber?: number;
 
    @Output() selectItem: EventEmitter<StTwoListSelectionElement> = new EventEmitter<StTwoListSelectionElement>();
    @Output() selectExtraLabel: EventEmitter<StTwoListSelectExtraLabelAction> = new EventEmitter<StTwoListSelectExtraLabelAction>();