+++ src/lib/st-two-list-selection/st-two-list-selection.model.ts
@@ -15,6 +15,7 @@ export interface StTwoListSelectionElement {
    name: string;
    selected?: boolean;
    extraLabel?: string;
+   itemAll?: boolean;
 }
 
 export interface StTwoListSelectExtraLabelAction {